+++ app-includes/search-edit-vestry.php
@@ -1,32 +1,34 @@
 <?php
 /**
- * This script displays the Search form for editing or deleting families. 
+ * This script displays the Search form for editing or deleting vestry members. 
  * It offers optional fields to narrow the search for duplicate last names. 
  * 
- * Upon 'SUBMIT' edit-family is run to process the input. 
+ * Upon 'SUBMIT' update-vestry is run to process the input. 
  */
 
 require_once __DIR__ . '/bootstrap.php';
+
+
 
 global $cota_db, $connect,  $cota_app_settings;
 require_once $cota_app_settings->COTA_APP_INCLUDES . 'headers.php';
 require_once $cota_app_settings->COTA_APP_INCLUDES . 'helper-functions.php';
 
 echo cota_page_header();
-$families = $cota_db->read_family_database();
-$num_families = $families->num_rows;
-if ( 0 == $num_families ) {
-	empty_database_alert('Search / Edit / Delete Directory Entries');
-    exit();
+$vestry_members = $cota_db->read_vestry_database();
+$num_vestry_members = $vestry_members->num_rows;
+if ( 0 == $num_vestry_members ) {
+	empty_database_alert('Search / Edit / Delete Vestry Entries');
+    // exit();
 } 
 // Grab a query parm if present. 
 $familyname = isset($_GET['familyname']) ? $_GET['familyname'] : '';
 
 // Dump out remainder of page.
 ?>
-    <h2>Search / Edit Family</h2>
-    <form class="cota-search" action="../app-includes/edit-family.php" method="get">
-        <label>Enter Family Name:</label>
+    <h2>Search / Edit Vestry Member</h2>
+    <form class="cota-search" action="../app-includes/update-vestry.php" method="get">
+        <label>Enter Vestry Member Family (Last) Name:</label>
         <input type="text" name="familyname" value="<?php echo htmlspecialchars($familyname); ?>">
         <p>OPTIONAL: The fields below may be used to differentiate families with same last names. </p>
         <label>Address</label>