+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.21] - 2026-06-03
+### Fixed
+- Removed deprecated function usage in RTF print script. Passing null to parameter #1 ($string) of type string in /var/www/cotadirectory/app-includes/class-membership-directory-printer.php on line 33
+
 ## [4.0.20] - 2026-06-03
 
 ### Changed